backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
openai==1.75.0
pymongo==4.11.3
python-dotenv==1.1.0
requests==2.32.3
